requirements.txt
jpmml-evaluator-pyspark
lightgbm
pyspark2pmml>=0.10.0
scikit-learn
sklearn2pmml